+++ src/file_service/mega/file_service/file_service_context.h
@@ -184,8 +184,9 @@ public:
     ~FileServiceContext();
 
     // Add a foreign file to the service.
-    auto add(NodeHandle handle, const common::NodeKeyData& keyData, std::size_t size)
-        -> FileServiceResultOr<FileID>;
+    auto add(NodeHandle handle,
+             const common::NodeKeyData& keyData,
+             std::uint64_t size) -> FileServiceResultOr<FileID>;
 
     // Notify observer when a file changes.
     FileEventObserverID addObserver(FileEventObserver observer);
